Issue 1: Done I think it's better if it's set from an option at the script level (the default behavior must be 0 to ask for a PIN). That option must also be applied to VA as well by setting the option data-nopin in the div myStart. And it can also be applied to authenticateExtended method with expectnopin parameter.
Issue 2: Done
On the push authenticate form could you add a link to the Manual OTP page
next to the Push button ?

Issue 3: Done
If user doesn't answer to the push notification the gluu server display a
HTTP 502 error. Instead a message should be displayed to the user on the
same form to say the notification expired but he can retry to connect. This
can be detected with error NOK:TIMEOUT

Issue 4: Done
I removed all the mobile phones associated to my account and asked for a
push, I'm redirected to the manual OTP form. Could you add an information
message on top of the form to explain the notification cannot be sent. This
issue can be detected with error NOK:NOPUSH and NOK:NOMA error from
pushAuthenticate

Issue 5:Done
If I enter a wrong username I have a Oops error from the server. This can be
detected with NOK:NOLOGIN error from pushAuthenticate